pyproject.toml
[project]
name = "pixel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.1.0",
    "requests>=2.32.3",
]
